refactor(reportes): Refactor reports module and add documentation
